backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt] 
email-validator
python-dotenv
supabase
bcrypt==4.0.1
pydantic
stripe
